+++ src/app/Domain/state/selected/selected.state.ts
@@ -1,0 +1,9 @@
+import { Selected } from '../../models/selected.model';
+
+export interface SelectedState {
+  selectedUser: Selected;
+}
+
+export const initialState: SelectedState = {
+  selectedUser: { name: '', testCase: -Infinity },
+}