+++ src/app/Domain/state/test-case/test-case.selector.ts
@@ -2,7 +2,7 @@ import { createFeatureSelector, createSelector } from '@ngrx/store';
 
 import { TestCaseState } from './test-case.state';
 
-export const TEST_CASE_NAME: string = 'test-case';
+export const TEST_CASE_NAME: string = 'testCase';
 
 const getTestCaseState = createFeatureSelector<TestCaseState>(TEST_CASE_NAME);