+++ src/app/Domain/state/warning/warning.state.ts
@@ -1,0 +1,9 @@
+import { Warning } from '../../models/warning.model';
+
+export interface WarningState {
+  warningsList: Warning[];
+}
+
+export const initialState: WarningState = {
+  warningsList: [],
+}